Several small bug fixes for device registration and equality comparisons.  Restructuring of folders to reduce noise and better organize the API
