+++ api/node/api/RegisterDevice/function.json
